requirements.txt
transformers==4.36.0
torch==2.1.0
newspaper3k==0.2.8
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
streamlit==1.28.0
pandas==2.1.3
numpy==1.24.3
nltk==3.8.1
urllib3==2.0.7
plotly==5.17.0
matplotlib==3.7.1
seaborn==0.12.2
jupyter==1.0.0
ipykernel==6.25.0
openpyxl==3.1.2

# New dependencies for enhanced features
feedparser==6.0.10
textstat==0.7.3
wordcloud==1.9.2
scikit-learn==1.3.2
python-dateutil==2.8.2
